# packages/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xiaohei-erp",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/bidding/rules": {
      "get": {
        "tags": [
          "bidding"
        ],
        "summary": "List Rules",
        "operationId": "list_rules_api_v1_bidding_rules_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Store Id"
            }
          },
          {
            "name": "sku",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sku"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingRuleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/stores/{store_id}/status": {
      "get": {
        "tags": [
          "bidding"
        ],
        "summary": "Store Status",
        "operationId": "store_status_api_v1_bidding_stores__store_id__status_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingStoreStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/stores/{store_id}/start": {
      "post": {
        "tags": [
          "bidding"
        ],
        "summary": "Start Store Bidding",
        "operationId": "start_store_bidding_api_v1_bidding_stores__store_id__start_post",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingStoreStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/stores/{store_id}/stop": {
      "post": {
        "tags": [
          "bidding"
        ],
        "summary": "Stop Store Bidding",
        "operationId": "stop_store_bidding_api_v1_bidding_stores__store_id__stop_post",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingStoreStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/stores/{store_id}/log": {
      "get": {
        "tags": [
          "bidding"
        ],
        "summary": "Store Log",
        "operationId": "store_log_api_v1_bidding_stores__store_id__log_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingRuleLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/stores/{store_id}/cycle": {
      "post": {
        "tags": [
          "bidding"
        ],
        "summary": "Run Cycle",
        "operationId": "run_cycle_api_v1_bidding_stores__store_id__cycle_post",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiddingCycleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingCycleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/rules/{rule_id}": {
      "patch": {
        "tags": [
          "bidding"
        ],
        "summary": "Update Rule",
        "operationId": "update_rule_api_v1_bidding_rules__rule_id__patch",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBiddingRuleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BiddingRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bidding/rules/bulk-import": {
      "post": {
        "tags": [
          "bidding"
        ],
        "summary": "Bulk Import Rules",
        "operationId": "bulk_import_rules_api_v1_bidding_rules_bulk_import_post",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Store Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BulkImportBiddingRuleItem"
                },
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportBiddingRuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/listing/jobs": {
      "get": {
        "tags": [
          "listing"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_listing_jobs_get",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Store Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingJobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listing/jobs/{job_id}": {
      "get": {
        "tags": [
          "listing"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_listing_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/listing/jobs/{job_id}/refresh-status": {
      "post": {
        "tags": [
          "listing"
        ],
        "summary": "Refresh Job Status",
        "operationId": "refresh_job_status_api_listing_jobs__job_id__refresh_status_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BiddingCycleRequest": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 50
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "BiddingCycleRequest"
      },
      "BiddingCycleResponse": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "real_write_enabled": {
            "type": "boolean",
            "title": "Real Write Enabled"
          },
          "processed_count": {
            "type": "integer",
            "title": "Processed Count"
          },
          "suggested_count": {
            "type": "integer",
            "title": "Suggested Count"
          },
          "applied_count": {
            "type": "integer",
            "title": "Applied Count"
          },
          "skipped_count": {
            "type": "integer",
            "title": "Skipped Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BiddingCycleItemResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "dry_run",
          "real_write_enabled",
          "processed_count",
          "suggested_count",
          "applied_count",
          "skipped_count",
          "failed_count",
          "items"
        ],
        "title": "BiddingCycleResponse"
      },
      "BiddingRuleListResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/BiddingRuleResponse"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "rules"
        ],
        "title": "BiddingRuleListResponse"
      },
      "BiddingRuleLogListResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/BiddingRuleResponse"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "rules"
        ],
        "title": "BiddingRuleLogListResponse"
      },
...
      "BiddingStoreStatusResponse": {
        "properties": {
          "store_id": {
            "type": "string",
            "title": "Store Id"
          },
          "is_running": {
            "type": "boolean",
            "title": "Is Running"
          },
          "active_rule_count": {
            "type": "integer",
            "title": "Active Rule Count"
          },
          "due_rule_count": {
            "type": "integer",
            "title": "Due Rule Count"
          },
          "blocked_count": {
            "type": "integer",
            "title": "Blocked Count"
          },
          "retrying_count": {
            "type": "integer",
            "title": "Retrying Count"
          },
          "fresh_count": {
            "type": "integer",
            "title": "Fresh Count"
          },
          "won_buybox_count": {
            "type": "integer",
            "title": "Won Buybox Count",
            "default": 0
          },
          "lost_buybox_count": {
            "type": "integer",
            "title": "Lost Buybox Count",
            "default": 0
          },
          "alert_count": {
            "type": "integer",
            "title": "Alert Count",
            "default": 0
          },
          "dry_run_default": {
            "type": "boolean",
            "title": "Dry Run Default"
          },
          "real_write_enabled": {
            "type": "boolean",
            "title": "Real Write Enabled"
          },
          "worker_enabled": {
            "type": "boolean",
            "title": "Worker Enabled"
          },
          "worker_cycle_limit": {
            "type": "integer",
            "title": "Worker Cycle Limit"
          },
          "last_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Started At"
          },
          "last_stopped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Stopped At"
          },
          "last_manual_cycle_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Manual Cycle At"
          },
          "last_worker_cycle_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Worker Cycle At"
          },
          "last_cycle_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Cycle Summary"
          }
        },
        "type": "object",
        "required": [
          "store_id",
          "is_running",
          "active_rule_count",
          "due_rule_count",
          "blocked_count",
          "retrying_count",
          "fresh_count",
          "dry_run_default",
          "real_write_enabled",
          "worker_enabled",
          "worker_cycle_limit"
        ],
        "title": "BiddingStoreStatusResponse"
      },
      "BulkImportBiddingRuleItem": {
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Sku"
          },
          "floor_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Price"
          }
        },
        "type": "object",
        "required": [
          "sku"
        ],
        "title": "BulkImportBiddingRuleItem"
      },
      "BulkImportBiddingRuleResponse": {
        "properties": {
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          },
          "created_count": {
            "type": "integer",
            "title": "Created Count"
          },
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/BiddingRuleResponse"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "imported_count",
          "created_count",
          "updated_count",
          "rules"
        ],
        "title": "BulkImportBiddingRuleResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListingJobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/ListingJobResponse"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "jobs"
        ],
        "title": "ListingJobListResponse"
      },
...
      "UpdateBiddingRuleRequest": {
        "properties": {
          "listing_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Id"
          },
          "floor_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Price"
          },
          "strategy_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "manual",
                  "guarded",
                  "aggressive"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Type"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "UpdateBiddingRuleRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}